% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_air.R
\name{airquality}
\alias{airquality}
\title{Extract data from pollution_us}
\usage{
airquality(data = pollution_us, pollutant, state, start, end)
}
\arguments{
\item{data}{a data.frame}

\item{pollutant}{type of pollutant (e.g. NO2, O3, SO2, CO)}

\item{state}{state name in US}

\item{start}{start date}

\item{end}{end date}
}
\description{
This function help you extract data needed.
}
