% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statemap.R
\name{checkAirQuality}
\alias{checkAirQuality}
\title{Check and plot air quality by state}
\usage{
checkAirQuality(data = AirQuality_Tracking, year)
}
\arguments{
\item{data}{A data.frame. The default dataset is AirQuality_Tracking.}

\item{year}{A numeric vector.}
}
\examples{
checkAirQuality(year=2010)

}
