% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_temp.R
\name{temp_state}
\alias{temp_state}
\title{Plot temperature geographic maps for States in USA}
\usage{
temp_state(data = GlobalLandTemperaturesByState, year)
}
\arguments{
\item{data}{A data.frame. The defalt dataset is GlobalLandTemperaturesByState.}

\item{year}{numeric}
}
\description{
This function plots temperature geographic maps for States in USA in specific year
}
\examples{
temp_state(year=2012)
select columns of interest
}
