% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Tidy Environment Dataset}
\usage{
tidy_data(data, date, na = "F")
}
\arguments{
\item{data}{a data.frame}

\item{date}{a character. It is the column name of the data and this column containing dates information.}

\item{na}{a character, either "F" or "T". The default value is "F". If the na=="F", it will remove rows with missing values.}
}
\value{
If na = "F", it will return a data frame without NA. The column names are with only lowercase letters, with "_" as a separator.
}
\description{
This function helps you tidy environment dataset, deal with the dates information, converts dates
    to Date class and create new column "Month" and "Year" to help you analyse the data.The column names are 
    with only lowercase letters, with "_" as a separator.
}
\examples{
tidy_data(pollution_us, "Date Local")

}
