% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend.R
\name{trend_aqi}
\alias{trend_aqi}
\title{Plot AQI trend}
\usage{
trend_aqi(data = pollution_us, city)
}
\arguments{
\item{data}{a data.frame}

\item{city}{city name in US}
}
\description{
This function plot NO2 AQI trend since 2000
}
