% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqi_healthconcern.R
\name{aqi_healthconcern}
\alias{aqi_healthconcern}
\title{Compare AQI category percentage by CBSA}
\usage{
aqi_healthconcern(data = annual_aqi, cbsa)
}
\arguments{
\item{data}{A data.frame. The default dataset is annual_aqi.}

\item{cbsa}{A vector of CBSA. You can input the CBSAs you want to compare.}
}
\description{
There are six categories of AQI. Each category corresponds to a different level of health concern. This function
    plots percentage of each category by CBSA. It can be used to compare AQI among different CBSAs.
}
\examples{
aqi_healthconcern(annual_aqi,c("Aberdeen, SD","Adrian, MI"))

}
