% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_temp.R
\name{avg_temp}
\alias{avg_temp}
\title{Average Monthly Temperatures Trend}
\usage{
avg_temp(data, year, month, country, con = "FALSE")
}
\arguments{
\item{data}{A data.frame. Use GlobalTemperatures or GlobalLandTemperaturesByCountry dataset to do analysis.}

\item{year}{A vector. Time period during which you want to analyse how the temperature changed.}

\item{month}{A vector. Time period during which you want to analyse how the temperature changed.}

\item{country}{A vector of countries' name. Compare how the temperature changed in these countries during a specified time period.}

\item{con}{Display confidence interval around smooth? (FALSE by default)}
}
\description{
This function helps you plot average monthly temperatures during a specified time period. You can also use it to do
    comparison between countries.
}
\examples{
avg_temp(data=GlobalLandTemperaturesByCountry,
year=c(2000:2015), month=c(1:12), 
country = c("Afghanistan","Andorra"),con="F")
}
