% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_aqi.R
\name{boxplot_aqi}
\alias{boxplot_aqi}
\title{Make the Boxplot of AQI}
\usage{
boxplot_aqi(data = annual_aqi, year, cbsa)
}
\arguments{
\item{data}{Dataset to use for plot. The default dataset is annual_aqi.}

\item{year}{A vector of years in which you want to display the distribution of AQI.}

\item{cbsa}{A CBSA in US.}
}
\description{
This function helps you make boxplots of AQI in the year you are interested. And also mark the point of the CBSA you are interested.
    So that you can know the AQI of the CBSA is relatively high or low compared to other geographic areas. You can make comparison of AQI in
    different years by the notch to see whether the AQI differs.
}
\examples{
boxplot_aqi(annual_aqi,c("2000","2005"),"Albany, GA")

}
\author{
Annie Yang
}
