% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_aqi.R
\name{boxplot_aqi}
\alias{boxplot_aqi}
\title{Make the boxplot of AQI}
\usage{
boxplot_aqi(data = annual_aqi, start, end, cbsa)
}
\arguments{
\item{data}{A data.frame. The default dataset is annual_aqi.}

\item{start}{Start year}

\item{end}{End year}

\item{cbsa}{A CBSA in US.}
}
\description{
This function helps you make boxplots of AQI in the year you are interested. And also mark the point of the CBSA you are interested.
}
\examples{
boxplot_aqi(annual_aqi,"2000","2005","Albany, GA")
}
