% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statemap.R
\name{checkAirQuality}
\alias{checkAirQuality}
\title{Check and plot air quality by state}
\usage{
checkAirQuality(data = AirQuality_Tracking, year)
}
\arguments{
\item{data}{Dataset to use for map. The default dataset is AirQuality_Tracking.}

\item{year}{A numeric. The report year of the air quality data.}
}
\description{
This function gets the data from CDC, the outcome is a combination outcome of
    monitoring ozone and PM2.5.
}
\examples{
checkAirQuality(year=2010)

}
\author{
Yiquan Xu
}
