% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_status.R
\name{health_status}
\alias{health_status}
\title{Check your health status}
\usage{
health_status(data = annual_aqi, cbsa, year, smoke = 1, exercise = 1,
  gene = 1)
}
\arguments{
\item{data}{a data.frame}

\item{cbsa}{cbsa in US}

\item{year}{A numeric vector. Time period during which you want to analyse the air quality.}

\item{smoke}{whether people smoke or not, equal to -1 if you smoke, else is 1. Default is 1}

\item{exercise}{whether people has certain exercise every day, equal to -1 if you don't, else is 1. Default is 1}

\item{gene}{whether people has certain gene may cause disease happen,equal to -1 for bad gene or 1 for good gene. Default is 1}
}
\description{
This function helps you to evaluate your health status
}
\examples{
health_status(cbsa = "Ames, IA", year = 2008)

}
