% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_func}
\alias{stat_func}
\title{Extract data from pollution_us}
\usage{
stat_func(data = annual_aqi_by_cbsa, cbsa)
}
\arguments{
\item{data}{a data.frame}

\item{cbsa}{cbsa in US}
}
\description{
This function help you extract data based on state, date and pollutant.
}
