% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_func}
\alias{stat_func}
\title{Calculate summary statistics for AQI}
\usage{
stat_func(data = annual_aqi, cbsa)
}
\arguments{
\item{data}{Dataset to use for summary statistics. The data is about the annual air quality index and other pollutants.}

\item{cbsa}{A CBSA in US.}
}
\value{
Return a data frame containing statistic information.
}
\description{
This function helps you calculate summary statistics for AQI by CBSA.
}
\examples{
stat_func(cbsa="Ames, IA")

}
\author{
Yiquan Xu
}
