% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_func}
\alias{stat_func}
\title{Gather statistical data from annual_aqi_by_cbsa}
\usage{
stat_func(data = annual_aqi_by_cbsa, cbsa)
}
\arguments{
\item{data}{a data.frame}

\item{cbsa}{cbsa in US}
}
\description{
This function helps you get the statistical data from annual_aqi_by_cbsa
}
