% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat_func}
\alias{stat_func}
\title{Calculate summary statistics for AQI}
\usage{
stat_func(data = annual_aqi, cbsa)
}
\arguments{
\item{data}{A data.frame}

\item{cbsa}{CBSA in US}
}
\description{
This function helps you calculate summary statistics for AQI by CBSA.
}
