% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_temp.R
\name{temp_country}
\alias{temp_country}
\title{Plot temperature geographic maps for countries}
\usage{
temp_country(data = GlobalLandTemperaturesByCountry, year)
}
\arguments{
\item{data}{A data.frame. The defalt dataset is GlobalLandTemperaturesByCountry.}

\item{year}{numeric}
}
\description{
This function plots temperature geographic maps for countries in specific year
}
\examples{
temp_country(year=2012)
}
