% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_aqi.R
\name{trend_aqi}
\alias{trend_aqi}
\title{Plot AQI trend}
\usage{
trend_aqi(data = pollution_us, city)
}
\arguments{
\item{data}{A data.frame}

\item{city}{US city name}
}
\description{
This function plots NO2, O3, SO2 and AQI trend in US city since 2000
}
\examples{
trend_aqi(pollution_us, "Phoenix")

}
